Fix: strip-animator: color cycle animations freezing
Some of the longer color cycle animations freeze because of an overflow
in the animation system. This overflow occurs because we are running the
animation system at a higher framerate than last year's badge.

Fortunately, we can now afford to store the tick count on a whole 16-bit
integer!

+++ components/badge-led-strip/strip_animator.hpp
@@ -193,7 +193,7 @@ class strip_animator : public scheduling::periodic_task<strip_animator>
         struct {
             union {
                 struct {
-                    uint8_t ticks_before_advance;
+                    uint16_t ticks_before_advance;
                     uint8_t star_count;
                 } shooting_star;
             };
@@ -218,7 +218,7 @@ class strip_animator : public scheduling::periodic_task<strip_animator>
 
             indice_storage_element origin_keyframe_index[8];
             indice_storage_element destination_keyframe_index[8];
-            uint8_t ticks_since_start_of_animation[16];
+            uint16_t ticks_since_start_of_animation[16];
         } keyframed;
     } _state;
     nsec::logging::logger _logger;
